[TEST] Do fewer tests per file speed up tests and prevent timeouts
