+++ test/integration/escape/invalid.test.js
@@ -1,0 +1,21 @@
+/**
+ * @overview Contains integration tests for invalid use of `shescape.escape`.
+ * @license MIT
+ */
+
+import { testProp } from "@fast-check/ava";
+import test from "ava";
+
+import { arbitrary, constants, macros } from "../_.js";
+
+import { escape as escape } from "shescape";
+
+testProp("invalid arguments", [arbitrary.shescapeOptions()], (t, options) => {
+  for (const { value } of constants.illegalArguments) {
+    t.throws(() => escape(value, options), { instanceOf: TypeError });
+  }
+});
+
+test(macros.prototypePollution, (_, payload) => {
+  escape("a", payload);
+});
