+++ test/integration/escape/powershell-exe.test.js
@@ -1,0 +1,21 @@
+/**
+ * @overview Contains integration tests for `shescape.escape` for Windows
+ * PowerShell.
+ * @license MIT
+ */
+
+import test from "ava";
+
+import { constants, generate } from "../_.js";
+
+import { escape as escape } from "shescape";
+
+const runTest = constants.isWindows ? test : test.skip;
+
+runTest(`input is escaped for ${constants.binPowerShell}`, (t) => {
+  for (const scenario of generate.escapeExamples(constants.binPowerShell)) {
+    const { expected, input, options } = scenario;
+    const result = escape(input, options);
+    t.is(result, expected);
+  }
+});
